Break SAML feature specs into multiple files
**Why**:
* Makes it easier to find specs
